// src/openapi/khronos/models/OverallStatsDto.ts
export interface OverallStatsDto {
    /**
     * Total number of bets
     * @type {number}
     * @memberof OverallStatsDto
     */
    totalBets: number;
    /**
     * Total number of wins
     * @type {number}
     * @memberof OverallStatsDto
     */
    totalWins: number;
    /**
     * Total number of losses
     * @type {number}
     * @memberof OverallStatsDto
     */
    totalLosses: number;
    /**
     * Win rate
     * @type {number}
     * @memberof OverallStatsDto
     */
    winRate: number;
    /**
     * Most bet team statistics
     * @type {TeamBetStats}
     * @memberof OverallStatsDto
     */
    mostBetTeam: TeamBetStats;
    /**
     * Team with most losses statistics
     * @type {TeamLossStats}
     * @memberof OverallStatsDto
     */
    mostLossesTeam: TeamLossStats;
    /**
     * Highest bet amount
     * @type {number}
     * @memberof OverallStatsDto
     */
    highestBetAmount: number;
    /**
     * Profit and loss summary
     * @type {ProfitLossSummary}
     * @memberof OverallStatsDto
     */
    profitLossSummary: ProfitLossSummary;
}